+++ model/wire/nnnsim/nnn/den/nnnsim-den.cc
@@ -98,7 +98,7 @@ namespace wire{
 
       poa_num = m_ptr->GetNumPoa ();
       // Get the serialized size from the Address class
-      for (int i = 0; i < poa_num; i++)
+      for (uint32_t i = 0; i < poa_num; i++)
 	{
 	  poatype_size += m_ptr->GetOnePoa (i).GetSerializedSize();
 	}
@@ -137,7 +137,7 @@ namespace wire{
       start.WriteU16(totalpoas);
 
       // Serialize PoAs
-      for (int i = 0; i < totalpoas; i++)
+      for (uint32_t i = 0; i < totalpoas; i++)
 	{
 	  Address tmpaddr = m_ptr->GetOnePoa(i);
 	  uint32_t serialSize = tmpaddr.GetSerializedSize ();
@@ -154,7 +154,7 @@ namespace wire{
 
 	  // Since the bit representation is in 8 bit chunks, serialize it
 	  // accordingly
-	  for (int j = 0; j < serialSize; j++)
+	  for (uint32_t j = 0; j < serialSize; j++)
 	    start.WriteU8(buffer[j]);
 	}
 
@@ -188,7 +188,7 @@ namespace wire{
 
       NS_LOG_INFO ("Deserialize -> PoA Num = " << totalpoas);
 
-      for (int k = 0; k < totalpoas; k++)
+      for (uint16_t k = 0; k < totalpoas; k++)
 	{
 	  uint8_t type = i.ReadU8 ();
 	  uint8_t length = i.ReadU8 ();
@@ -196,7 +196,7 @@ namespace wire{
 	  // Create a buffer to be able to deserialize PoAs
 	  uint8_t buffer[length];
 
-	  for (int j = 0; j < length; j++)
+	  for (uint8_t j = 0; j < length; j++)
 	    {
 	      buffer[j] = i.ReadU8 ();
 	    }
